% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiniertreecheckbox.R
\name{shiniertreecheckbox}
\alias{shiniertreecheckbox}
\title{ShinierTreeCheckbox}
\usage{
shiniertreecheckbox(elementId, data, options = NULL)
}
\arguments{
\item{elementId}{The ID of the HTML element used to hold the tree.}

\item{data}{A list of nodes to display in the tree.}

\item{options}{A list of options for the tree.}
}
\description{
See the Github page for more information: https://github.com/SanderJBouwman/shiniertreecheckbox
}
